refactor: nats new structure
