+++ faststream/nats/schemas/__init__.py
@@ -1,0 +1,7 @@
+from faststream.nats.schemas.js_stream import JStream
+from faststream.nats.schemas.pull_sub import PullSub
+
+__all__ = (
+    "JStream",
+    "PullSub",
+)
