+++ faststream/nats/schemas/js_stream.py
