+++ faststream/nats/schemas/js_stream.pyi
